requirements.txt
spacy
streamlit
google-generativeai
python-dotenv
python-docx